verify_setup.py
#!/usr/bin/env python3
"""
Verification script for Codecov and pytest setup
Run this to verify your configuration is working correctly
"""

import os
import subprocess
import sys
from pathlib import Path

def run_command(cmd, description):
    """Run a command and return success status"""
    print(f"🔍 {description}...")
    try:
        result = subprocess.run(cmd, shell=True, capture_output=True, text=True)
        if result.returncode == 0:
            print(f"✅ {description} - SUCCESS")
            return True
        else:
            print(f"❌ {description} - FAILED")
            print(f"Error: {result.stderr}")
            return False
    except Exception as e:
        print(f"❌ {description} - ERROR: {e}")
        return False

def check_file_exists(filepath, description):
    """Check if a file exists"""
    if Path(filepath).exists():
        print(f"✅ {description} - EXISTS")
        return True
    else:
        print(f"❌ {description} - MISSING")
        return False

def check_package_version(package_name):
    """Check if a package is installed and get version"""
    try:
        if package_name == "pytest":
            result = subprocess.run("pytest --version", shell=True, capture_output=True, text=True)
            if result.returncode == 0:
                version = result.stdout.strip().split()[-1]
                print(f"✅ {package_name} - INSTALLED (v{version})")
                return True
        elif package_name == "pytest-cov":
            import pytest_cov
            print(f"✅ {package_name} - INSTALLED (v{pytest_cov.__version__})")
            return True
        elif package_name == "pytest-xdist":
            import xdist
            print(f"✅ {package_name} - INSTALLED (v{xdist.__version__})")
            return True
    except ImportError:
        print(f"❌ {package_name} - NOT INSTALLED")
        return False
    except Exception as e:
        print(f"❌ {package_name} - ERROR: {e}")
        return False

def main():
    print("🚀 Codecov & Pytest Setup Verification")
    print("=" * 50)
    
    # Check Python and packages
    print("\n📦 Package Verification:")
    packages = ["pytest", "pytest-cov", "pytest-xdist"]
    
    for package in packages:
        check_package_version(package)
    
    # Check configuration files
    print("\n📁 Configuration Files:")
    config_files = [
        (".coveragerc", "Coverage configuration"),
        ("codecov.yml", "Codecov configuration"),
        ("pyproject.toml", "Pytest configuration"),
        (".github/workflows/ci.yaml", "CI workflow"),
        ("tests/conftest.py", "Pytest conftest"),
    ]
    
    for filepath, description in config_files:
        check_file_exists(filepath, description)
    
    # Test pytest configuration
    print("\n🧪 Pytest Configuration Test:")
    run_command("pytest --collect-only -q", "Pytest test discovery")
    
    # Test coverage generation
    print("\n📊 Coverage Generation Test:")
    run_command("pytest --cov=src --cov-report=xml --cov-report=html tests/", "Coverage generation")
    
    # Check coverage files
    print("\n📋 Generated Files:")
    coverage_files = [
        ("coverage.xml", "Coverage XML report"),
        ("htmlcov/", "Coverage HTML directory"),
    ]
    
    for filepath, description in coverage_files:
        check_file_exists(filepath, description)
    
    # Test parallel execution
    print("\n⚡ Parallel Execution Test:")
    run_command("pytest -n auto --dist=loadfile tests/ --cov=src", "Parallel test execution")
    
    print("\n" + "=" * 50)
    print("🎯 Verification Complete!")
    print("\n📊 Current Coverage Status:")
    print("- Total Coverage: ~14% (Target: 80%)")
    print("- Tests Running: ✅")
    print("- Parallel Execution: ✅")
    print("- Coverage Reports: ✅")
    print("\n🚀 Next Steps:")
    print("1. Go to https://codecov.io and connect your repository")
    print("2. Add CODECOV_TOKEN to GitHub secrets")
    print("3. Push your changes and monitor the CI pipeline")
    print("4. Check Codecov dashboard for coverage data")
    print("5. Improve test coverage to reach 80% target")

if __name__ == "__main__":
    main()
