requirements.txt
evidently==0.7.12
fastapi==0.116.1
joblib==1.4.2
mlflow==2.17.2
numpy==1.24.4
pandas==2.0.3
pydantic==2.10.6
PyYAML==6.0.2
scikit-learn==1.3.2
xgboost==2.1.4
urllib3==2.2.3
pytest-xdist>=3.0.0
pytest-cov>=4.0.0